Refactor cli call out of parameter generators
